1373 DR History
Added Sammaster and his history.
Sammaster rules over Vaasa briefly and invades Bloodstone.
Ylraphon is destroyed in the Rage.
Tilverton is resettled.
Tangled Trees is destroyed in the Rage.
Rawlinswood is "freed" from the Blightlords by a Demon Horde.
Tchazzar takes over Cimbar and the crown of Chessenta.
Corrected a minor issue.