Empire-tier Phalorm
Phalorm is now empire tier, with Delimbiyran, Ardeep and Dardath as constituent kingdoms; Amristar and Greyvale (to be changed to Glorea later) are constituent duchies; Navaraemista, Hunnabar, Scathril, and Calantor are subjects of the culturally-appropriate kingdoms.

Also, Narthil (Misty Forest) now falls to Athalantar, but Navaraemista survives until the end of Phalorm. Pharren reverts to Eaerlann after the fall of Athalantar.