cleanslate updates
The Matriarchal Deposition CB no longer sets enatic succession if it will reset on the next succession, due to being invalid. It now tries to set as female-centric succession as the involved titles will allow for. The Patriarchal Deposition CB does the reverse, applying a maximally possible male-centric succession law